requirements.txt
Django>=4.0,<5.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
psycopg2>=2.9.0
gunicorn>=20.1.0
pybreaker>=1.2.0